// src/graphql/types.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A string containing HTML code. */
  HTML: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type FollowerConnection = {
  __typename?: "FollowerConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<UserEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars["String"]>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars["Boolean"];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars["Boolean"];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars["String"]>;
};
// ...
export type StarredRepositoryConnection = {
  __typename?: "StarredRepositoryConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<StarredRepositoryEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Repository>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};

/** Represents a starred repository. */
export type StarredRepositoryEdge = {
  __typename?: "StarredRepositoryEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  node: Repository;
  /** Identifies when the item was starred. */
  starredAt: Scalars["DateTime"];
};
// ...
export type UserEdge = {
  __typename?: "UserEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of the edge. */
  node?: Maybe<User>;
};
// ...
export type UserFragment = { __typename?: "User" } & Pick<
  User,
  "id" | "login" | "name" | "avatarUrl" | "bio" | "email" | "location"
> & {
    starredRepositories: { __typename?: "StarredRepositoryConnection" } & Pick<
      StarredRepositoryConnection,
      "totalCount"
    >;
    followers: { __typename?: "FollowerConnection" } & Pick<
      FollowerConnection,
      "totalCount"
    >;
  };